backend/pyproject.toml
[project]
name = "ai-news"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "feedparser>=6.0.12",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=25.1.0",
    "jinja2>=3.1.6",
    "litellm>=1.82.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
